requirements.txt
fastapi
aiomysql
python-jose
sqlalchemy

